Add programs written by @cpsdqs to display a few number of the fibonacci sequence.
These are pretty complex programs for a toy ISA like this one. It requires implementing a stack for recursion, and in the case of the ASCII version, a routine for outputting a number to the screen without division. Both of these eat up tons of CPU cycles.
You can see the calling convention at the top of the file.